+++ src/core/widget_store.rs
@@ -97,13 +97,17 @@ impl WidgetStore {
     /// After adding a widget, the ID of the widget is returned.
     pub fn add_widget(&mut self, name: &str, widget: Box<dyn Widget>) -> i32 {
         let widget_size = self.widgets.len() as i32;
-
-        self.widgets.push(WidgetContainer {
+        let container = WidgetContainer {
             widget: RefCell::new(widget),
             widget_name: String::from(name),
             widget_id: widget_size,
             parent_id: 0,
-        });
+        };
+
+        // #117 - assigns widget ID to itself
+        container.widget.borrow_mut().config().set_numeric(CONFIG_WIDGET_ID, widget_size as u64);
+
+        self.widgets.push(container);
 
         widget_size
     }
@@ -120,13 +124,17 @@ impl WidgetStore {
     ) -> i32 {
         // TODO Validate parent_id
         let widget_size = self.widgets.len() as i32;
-
-        self.widgets.push(WidgetContainer {
+        let container = WidgetContainer {
             widget: RefCell::new(widget),
             widget_name: String::from(name),
             widget_id: widget_size,
             parent_id,
-        });
+        };
+
+        // #117 - assigns widget ID to itself
+        container.widget.borrow_mut().config().set_numeric(CONFIG_WIDGET_ID, widget_size as u64);
+
+        self.widgets.push(container);
 
         widget_size
     }